Move in_projection to separate file
